Don't build man pages for OTP 27

+++ .github/scripts/release-docs.sh
@@ -1,8 +1,14 @@
-#!/bin/sh
+#!/bin/bash
 
-docker run -v $PWD/:/github otp "make release docs release_docs && make release_docs DOC_TARGETS='man html pdf' RELEASE_ROOT=/github/docs"
+if [[ $1 < 27 ]]; then
+    docker run -v $PWD/:/github otp "make release docs release_docs && make release_docs DOC_TARGETS='man html pdf' RELEASE_ROOT=/github/docs"
+else
+    docker run -v $PWD/:/github otp ".otp_build download_ex_doc && make release docs release_docs && make release_docs DOC_TARGETS=html RELEASE_ROOT=/github/docs"
+fi
 sudo chown -R `whoami` docs
 cd docs
-tar czf ../otp_doc_man.tar.gz man
-rm -rf man
+if [[ $1 < 27 ]]; then
+    tar czf ../otp_doc_man.tar.gz man
+    rm -rf man
+fi
 tar czf ../otp_doc_html.tar.gz *
